# docs/swagger.yaml
basePath: /
definitions:
...
  admin.createMovieTypeRequest:
    properties:
      name:
        type: string
    type: object
...
  models.ApiError:
    properties:
      error:
        type: string
    type: object
...
  models.Genre:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
...
  models.MovieType:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: bexultanovmadiyar@gmail.com
    name: Madiyar Bexultanov
    url: https://www.linkedin.com/in/madiyar-bexultanov-b21902258/
  description: Personal online platform providing information about media content
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ozinshe Production
  version: "1.0"
paths:
...
  /admin/genres:
    get:
      consumes:
      - application/json
      description: Retrieve all genres from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Genre'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all genres
      tags:
      - genres
    post:
      consumes:
      - multipart/form-data
      description: Create a new genre and upload a poster
      parameters:
      - description: Title of the genre
        in: formData
        name: title
        required: true
        type: string
      - description: Poster image
        in: formData
        name: poster
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: id of the created genre
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create a new genre
      tags:
      - genres
  /admin/genres/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a genre from the database by its ID
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Genre deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete a genre by ID
      tags:
      - genres
    get:
      consumes:
      - application/json
      description: Retrieve a genre by its ID
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Genre'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get genre by ID
      tags:
      - genres
    put:
      consumes:
      - application/json
      description: Update a genre's details, including its title and poster
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated genre data
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/models.Genre'
      produces:
      - application/json
      responses:
        "200":
          description: Genre updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update a genre by ID
      tags:
      - genres
  /admin/movieTypes:
    get:
      consumes:
      - application/json
      description: Get a list of all movie types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MovieType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all movie types
      tags:
      - MovieTypes
    post:
      consumes:
      - application/json
      description: Create a new movie type with the specified name
      parameters:
      - description: Movie Type Information
        in: body
        name: movieType
        required: true
        schema:
          $ref: '#/definitions/admin.createMovieTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Movie Type ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create a new movie type
      tags:
      - MovieTypes
  /admin/movieTypes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a movie type by its ID
      parameters:
      - description: Movie Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Movie type deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete a movie type
      tags:
      - MovieTypes
    get:
      consumes:
      - application/json
      description: Get a movie type by its ID
      parameters:
      - description: Movie Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MovieType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get movie type by ID
      tags:
      - MovieTypes
    put:
      consumes:
      - application/json
      description: Update an existing movie type by its ID
      parameters:
      - description: Movie Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Movie Type Information
        in: body
        name: movieType
        required: true
        schema:
          $ref: '#/definitions/models.MovieType'
      produces:
      - application/json
      responses:
        "200":
          description: Movie type updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update an existing movie type
      tags:
      - MovieTypes
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"